update ip controller unit test;
