+++ src/routes/index.js
@@ -1,7 +1,5 @@
 import express from 'express';
 
-import AdminRoutes from '#src/routes/admin.route.js';
-import UserRoutes from '#src/routes/user.route.js';
 import FabRoutes from '#src/routes/fab.route.js';
 import IpRoutes from '#src/routes/ip.route.js';
 import RoomRoutes from '#src/routes/room.route.js';
@@ -14,8 +12,6 @@ router.get('/healthy', (req, res) => {
   res.send('Backend service is healthy.');
 });
 
-router.use('/admin', AdminRoutes);
-router.use('/user', UserRoutes);
 router.use('/fab', FabRoutes);
 router.use('/ip', IpRoutes);
 router.use('/room', RoomRoutes);
