+++ src/routes/ip.route.js
@@ -1,7 +1,7 @@
 import express from 'express';
 
 import {
-  assignIpController,
+  assignController,
   createIpPoolController,
   releaseController,
   getAllIpController,
@@ -10,7 +10,7 @@ import {
 
 const router = express.Router();
 
-router.post('/', assignIpController); // Assign an IP address to a fabId and service
+router.post('/', assignController); // Assign an IP address to a fabId and service
 router.post('/pool', createIpPoolController); // Create an IP pool for a fabId and service
 router.delete('/', releaseController); // release an IP address for a fabId and service
 router.get('/allIp', getAllIpController); // Get all IPs for a fabId and service
